# .skmtc/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Pet store - Dogs",
    "version": "0.0.0"
  },
  "tags": [
    {
      "name": "Dogs"
    }
  ],
  "paths": {
    "/dogs": {
      "get": {
        "operationId": "Dogs_list",
        "summary": "Get all dogs",
        "description": "List of all dogs in store",
        "parameters": [
          {
            "name": "name",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Filter results by name"
            },
            "explode": false
          }
        ],
        "responses": {
          "200": {
            "description": "The request has succeeded.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DogList"
                }
              }
            }
          }
        },
        "tags": [
          "Dogs"
        ]
      },
      "post": {
        "operationId": "Dogs_create",
        "summary": "Create new dog",
        "description": "Add a new dog to store.",
        "parameters": [],
        "responses": {
          "200": {
            "description": "The request has succeeded.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Dog"
                }
              }
            }
          }
        },
        "tags": [
          "Dogs"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Dog"
              }
            }
          }
        }
      },
      "put": {
        "operationId": "Dogs_update",
        "summary": "Update dog",
        "description": "Add a new dog to store.",
        "parameters": [],
        "responses": {
          "200": {
            "description": "The request has succeeded.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Dog"
                }
              }
            }
          }
        },
        "tags": [
          "Dogs"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Dog"
              }
            }
          }
        }
      }
    },
    "/dogs/{id}": {
      "get": {
        "operationId": "Dogs_get",
        "summary": "Get dog by id",
        "description": "Get a dog by id",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The request has succeeded.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Dog"
                }
              }
            }
          }
        },
        "tags": [
          "Dogs"
        ]
      }
    }
  },
  "components": {
    "schemas": {
      "Breed": {
        "type": "string",
        "enum": [
          "dachshund",
          "dalmatian",
          "poodle"
        ]
      },
      "Dog": {
        "type": "object",
        "required": [
          "id",
          "name",
          "breed"
        ],
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "breed": {
            "$ref": "#/components/schemas/Breed"
          }
        },
        "example": {
          "id": "1",
          "name": "Archie",
          "breed": "dachshund"
        }
      },
      "DogList": {
        "type": "array",
        "items": {
          "$ref": "#/components/schemas/Dog"
        },
        "example": [
          {
            "id": "1",
            "name": "Archie",
            "breed": "dachshund"
          },
          {
            "id": "2",
            "name": "Amos",
            "breed": "dachshund"
          }
        ]
      }
    }
  },
  "servers": [
    {
      "url": "https://example.com",
      "variables": {}
    }
  ]
}